generalize settings application
