+++ module/damage/settings.ts
@@ -1,3 +1,4 @@
+import { GurpsSettingsApplication } from '../utilities/gurps-settings-application.js'
 import {
   APPLY_DIVISOR,
   BLUNT_TRAUMA,
@@ -276,77 +277,8 @@ export function defaultADDAction(): string {
   return game.settings?.get(GURPS.SYSTEM_NAME, DEFAULT_ADD_ACTION) ?? DEFAULT_ADD_ACTION
 }
 
-class DamageSettingsApplication extends foundry.applications.api.HandlebarsApplicationMixin(
-  foundry.applications.api.ApplicationV2
-) {
+class DamageSettingsApplication extends GurpsSettingsApplication {
   constructor(options?: any) {
-    super(options)
-  }
-
-  static override DEFAULT_OPTIONS = {
-    classes: ['gga', 'standard-form'],
-    form: {
-      closeOnSubmit: true,
-      handler: DamageSettingsApplication.update,
-    },
-    id: 'gga-settings',
-    position: {
-      width: 600,
-      height: 600,
-    },
-    tag: 'form',
-    window: {
-      resizable: true,
-      icon: 'fa-light fa-face-head-bandage',
-    },
-  }
-
-  static override PARTS = {
-    main: {
-      scrollable: ['settings-list'],
-      template: 'systems/gurps/templates/damage/settings.hbs',
-    },
-  }
-
-  override get title() {
-    return game.i18n!.localize(SETTINGS)
-  }
-
-  protected override async _prepareContext(
-    options: foundry.applications.api.ApplicationV2.RenderOptions & { isFirstRender: boolean }
-  ): Promise<foundry.applications.api.ApplicationV2.RenderContext> {
-    const context = await super._prepareContext(options)
-
-    const settings =
-      (Array.from(game.settings!.settings.values()).filter((s: any) => s.id.startsWith('gurps.damage')) as any) || []
-
-    // go through all settings, and collect their values
-    for (const setting of settings) setting.value = game.settings!.get(GURPS.SYSTEM_NAME, setting.key as any)
-
-    const result = foundry.utils.mergeObject(context, {
-      settings: settings,
-    })
-    return result
-  }
-
-  static async update(event: SubmitEvent | Event, form: HTMLFormElement, formData: FormDataExtended): Promise<void> {
-    event.preventDefault()
-    event.stopPropagation()
-
-    const data = foundry.utils.expandObject(formData) as Record<string, any>
-
-    data
-      .keys()
-      .toArray()
-      .forEach(async (key: any) => {
-        const namespace = key.split('.')[0]
-        const settingId = key.split('.').slice(1).join('.')
-        if (namespace !== GURPS.SYSTEM_NAME) {
-          console.warn(`GURPS | DamageSettingsApplication.update: Skipping setting ${key} with namespace ${namespace}`)
-          return
-        }
-
-        await game.settings!.set(GURPS.SYSTEM_NAME, settingId, data.object[key])
-      })
+    super({ title: game.i18n!.localize(SETTINGS), module: MODULE_NAME }, options)
   }
 }
